stop drawing partial squares on ends of rows
unless the width is an exact multiple of 'cell size', you ended up with
partial squares on the end of rows. this commit fixes the drawing logic to
only add a square to the end of a row if there is enough width for the
whole of it to be drawn, otherwise it moves it to the next line.

fixes #25