fix misaligned positioning of squares
borders overlap squares with half of their width, thus squares need a
surrounding margin of half the configured border width. the squares in the
first row and column were being drawn right up against the edges of the
drawing area without leaving such room for the border.

this resulted in two noticeable issues:
 - firstly, half of the left and top borders of the squares in the leftmost
   row and topmost column respectively was hidden, which looked odd.
 - secondly, this problematic positioning of the squares did not meet the
   expectations of the selection code that determined which square you are
   hovering over or clicking on, such that the selection areas were
   misaligned. rather than a square's selection areas covering the square
   and half of the surrounding space on each side, it instead was covering
   the square and essentially all space to the right and bottom, and none
   to the top and left.

offsetting the drawing of the squares to correctly account for the borders
fixes both of these issues.

fixes #30, #31