fix resizing upon switching selected torrent
if you select a torrent with a big number of pieces, then switch to one
with fewer, the area in which the pieces are drawn could remain the
previous size (it grew but never shrank), which results in both a vertical
scroll bar when one is unnecessary and the priority checkbox being located
much further down than it should be.

calculating the number of rows and directly setting the needed height up
front works better than adding an additional call to set the size after the
loop, since this makes fewer calls and `rows` is useful to the next commit.

fixes #29